画像リサイズのUIImageのextensionをViewControllerファイルから分離
